test: split test_app_lifecycle_service_operations.py by operation family
The file was 1761 lines — the largest test file failing the house-lint
file-size check (HSL102, 800-line threshold). Its 24 test classes had clean
boundaries, so it splits into four sibling files by operation family:

- test_app_lifecycle_service_start_stop.py (440) — whole-app start/stop
- test_app_lifecycle_service_reload.py (560) — apply_changes and reload
- test_app_lifecycle_service_instance_ops.py (575) — per-instance ops/locking
- test_app_lifecycle_service_reconcile.py (216) — reconcile/config/manifests

Class bodies are byte-identical to the original; only imports were re-derived
per file. Test count is unchanged (67 in the split files, 140 across
tests/unit/core/test_app_lifecycle*.py).

Closes #1756
